requirements.txt
-i https://pypi.org/simple
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
ffmpeg-python==0.2.0
flask==2.2.2
flask-cors==3.0.10
future==0.18.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.2
numpy==1.24.1
opencv-python==4.7.0.68
pydantic==1.10.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
werkzeug==2.2.2
